% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.character.czech_date}
\alias{as.character.czech_date}
\title{as.character S3 method for class czech_date}
\usage{
\method{as.character}{czech_date}(date, ..., case = NULL)
}
\description{
as.character S3 method for class czech_date
}
\keyword{internal}
