% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{b64_string_to_file}
\alias{b64_string_to_file}
\title{Write base64 string to file}
\usage{
b64_string_to_file(b64_string, output_file)
}
\arguments{
\item{b64_string}{\emph{character}, base64-encoded string to decode and write to file}

\item{output_file}{\emph{character}, path to the output file where the decoded content will be written}
}
\value{
side effect of writing a file
}
\description{
Write base64 string to file
}
