% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clr2hex}
\alias{clr2hex}
\title{Convert color to hexadecimal format}
\usage{
clr2hex(color, alpha = 1)
}
\arguments{
\item{color}{\emph{character}, color name (seel \code{colors()} for recognized values)}

\item{alpha}{\emph{numeric}, opacity in interval 0--1, where 1 is no transparency,
i.e. full opacity}
}
\value{
\emph{character}, color in hexadecimal format
}
\description{
Convert color to hexadecimal format
}
\keyword{internal}
