% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_schola_template}
\alias{copy_schola_template}
\title{Copy default Schola template into project directory}
\usage{
copy_schola_template(format = "pdf", path = proj_get(), ...)
}
\arguments{
\item{format}{\emph{Character}, format which to look for. Defaults to \code{pdf}.}

\item{path}{\emph{Character}, path to copy to. Defaults to the current project
root.}

\item{...}{
  Arguments passed on to \code{\link[base:files]{base::file.copy}}
  \describe{
    \item{\code{overwrite}}{logical; should existing destination files be overwritten?}
    \item{\code{recursive}}{logical.  If \code{to} is a directory, should
    directories in \code{from} be copied (and their contents)?  (Like
    \command{cp -R} on POSIX OSes.)}
    \item{\code{copy.mode}}{logical:  should file permission bits be copied where
    possible?}
    \item{\code{copy.date}}{logical:  should file dates be preserved where
    possible? See \code{\link[base]{Sys.setFileTime}}.}
  }}
}
\value{
No return value, called for side effect.
}
\description{
Copy default Schola template into project directory
}
\examples{
\dontrun{
copy_schola_template()
}
}
\seealso{
Other Report templates and formats: 
\code{\link{open_schola_template}()},
\code{\link{schola_pdf}()},
\code{\link{schola_word}()},
\code{\link{schola_word2}()}
}
\author{
Jan Netik
}
\concept{Report templates and formats}
