% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_word}
\alias{create_word}
\title{Create a Rmd doc using schola_word template}
\usage{
create_word(name = "draft.Rmd", open = T)
}
\arguments{
\item{name}{name to use for new file.}

\item{open}{whether to open file for editing, defaults to TRUE.}
}
\description{
Shortcut function to create and a new Rmd file using schola_word template and open for editing.
}
\examples{
\dontrun{
  create_word("new_file")
}
}
