% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{czech_date_interval}
\alias{czech_date_interval}
\title{Czech Date Interval}
\usage{
czech_date_interval(start, end)
}
\arguments{
\item{start}{\emph{Date of date-like object}, start date or left boundary of an
interval.}

\item{end}{\emph{Date of date-like object}, end date or right boundary of an
interval.}
}
\value{
Character
}
\description{
Returns the most space-efficient and at the same time grammatically correct
interval of two Czech dates. When both dates are the same, only one is
outputted. The function ensures that the interval is not negative (i.e.,
\code{start} <= \code{end}), otherwise, it is reversed.
}
\examples{
czech_date_interval("2020-01-24", "2020-01-03") # note the argument order
}
