% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{czech_date_main}
\alias{czech_date_main}
\title{Czech Date Internals}
\usage{
czech_date_main(date, case = "genitive")
}
\arguments{
\item{date}{\emph{date or date-like object} to parse.}

\item{case}{\emph{character}, either "nominative", "locative" or "genitive"
(default) or any unambiguous abbreviation of these.}
}
\description{
Function used by S3 methods for class \code{czech_date}.
}
\section{Grammatical cases}{

Three grammatical cases are supported:
\itemize{
\item \emph{nominative} -- native form, i.e. "leden" in Czech
\item \emph{locative} -- "in ...", i.e. "v lednu" in Czech
\item \emph{genitive} -- "the 'nth' of ...", i.e. "5. ledna" in Czech
}

Czech months listed by case are available in \code{.czech_months}.
}

\keyword{internal}
