% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{dict_from_data}
\alias{dict_from_data}
\title{Get item code-label dictionary from data}
\usage{
dict_from_data(.data)
}
\arguments{
\item{.data}{data.frame or tibble}
}
\value{
named character vector of items' labels, if no label found, NULL is
introduced and is omitted in the output
}
\description{
Item labels have to be stored as variable attribute \code{label}. Exported by
default by \code{ls_export()}, but many data wrangling operations strip them out,
unfortunately.
}
