% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diverging-chart.R
\name{diverging_chart}
\alias{diverging_chart}
\title{FUNCTION_TITLE}
\usage{
diverging_chart(data)
}
\arguments{
\item{data}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
