% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.czech_months}
\alias{.czech_months}
\title{List of Czech Months in Three Grammatical Cases}
\format{
An object of class \code{list} of length 3.
}
\usage{
.czech_months
}
\description{
As used by \code{czech_date_internal()}.
}
\keyword{internal}
