% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{draft_pdf}
\alias{draft_pdf}
\alias{draft_word}
\title{Create a \code{.Rmd} draft using Schola templates}
\usage{
draft_pdf(name = "pdf_draft", open = TRUE)

draft_word(name = "word_draft", open = TRUE)
}
\arguments{
\item{name}{\emph{character}, name to use for new file. With or without file extension.}

\item{open}{\emph{logical}, whether to open file for editing, defaults to TRUE.}
}
\value{
Path to the created file (invisibly).
}
\description{
Shortcut function to create a new \code{.Rmd} file using Schola standard templates
and open it for editing.
}
\examples{
\dontrun{
draft_pdf("best_report")
}

}
\seealso{
Other Workflow helpers: 
\code{\link{gd_download_folder}()},
\code{\link{manage_docx_header_logos}()}
}
\concept{Workflow helpers}
