% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{draft_redoc}
\alias{draft_redoc}
\title{Create a Rmd doc for reversible reviewing using schola_redoc template}
\usage{
draft_redoc(name = "draft.Rmd", open = T)
}
\arguments{
\item{name}{name to use for new file.}

\item{open}{whether to open file for editing, defaults to TRUE.}
}
\description{
Shortcut function to create and a new Rmd file using schola_redoc template and open for editing.
}
\examples{
\dontrun{
draft_word("new_file")
}
}
