% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{draft_word}
\alias{draft_word}
\title{Create a Rmd doc using schola_word template}
\usage{
draft_word(name = "draft.Rmd", open = T)
}
\arguments{
\item{name}{name to use for new file.}

\item{open}{whether to open file for editing, defaults to TRUE.}
}
\value{
path to created file (invisibly)
}
\description{
Shortcut function to create and a new Rmd file using schola_word template and open for editing.
}
\examples{
\donttest{
draft_word("new_file")
}
}
\seealso{
Other Workflow helpers: 
\code{\link{draft_redoc}()},
\code{\link{gd_download_folder}()},
\code{\link{manage_docx_header_logos}()}
}
\concept{Workflow helpers}
