% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_output.R
\name{ensure_cropping}
\alias{ensure_cropping}
\title{Ensure That Figure Cropping Dependencies Are Satisfied}
\usage{
ensure_cropping()
}
\value{
No return value. Called for side effects.
}
\description{
Checks if the dependencies needed for the \code{schola_pdf} format to work
properly and as intended are satisfied. If not, solution is offered.
}
\details{
In order to use automated cropping of PDF figures, the \code{PDFCrop} utility
(shipped with TinyTeX, MiKTeX, or TeX Live distributions) is required. The
program further depends on \code{GhostScript}, which is usually not set on PATH
upon installation and you have to do it manually. The good news is that it
can be done straight from the R.
}
