% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_schola_barplot_info}
\alias{extract_schola_barplot_info}
\title{Extract detailed info from Schola barplot}
\usage{
extract_schola_barplot_info(plot)
}
\arguments{
\item{plot}{result of schola_barplot() call}
}
\value{
tibble with plot data
}
\description{
Note that item labels are automatically derived using the labeller from given \code{schola_barplot} result
}
\examples{
"TODO"
}
