% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{fct_nanify}
\alias{fct_nanify}
\title{NAnify factor level}
\usage{
fct_nanify(f, level)
}
\arguments{
\item{f}{\emph{factor} to work on}

\item{level}{\emph{character},}
}
\value{
\emph{factor} with NA-substituted level.
}
\description{
The inverse of \code{fct_explicit_na()}. Turns the \code{level} into proper \code{NA}.
\strong{Retains the level.}
}
\examples{
f <- factor(c("a", "b", "c", "nanify"))
fct_nanify(f, "nanify")
}
