% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{flush_axis}
\alias{flush_axis}
\title{A shorcut for making axis text flush with axis}
\format{
An object of class \code{numeric} of length 4.
}
\usage{
flush_axis
}
\description{
A shorcut for making axis text flush with axis
}
\examples{
library(ggplot2)
ggplot(mpg) +
  geom_bar(aes(y = class)) +
  scale_x_continuous(expand = flush_axis)
}
\seealso{
Other Making charts: 
\code{\link{plot_lollipop}()},
\code{\link{prepare_lollipop_data}()},
\code{\link{schola_barplot}()},
\code{\link{theme_schola}()}
}
\concept{Making charts}
\keyword{datasets}
