% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{formatscz}
\alias{formatscz}
\alias{label_percent_cz}
\alias{label_number_cz}
\title{Czech formatted scales for ggplot2}
\usage{
label_percent_cz(
  accuracy = NULL,
  scale = 100,
  prefix = "",
  suffix = " \%",
  big.mark = " ",
  decimal.mark = ",",
  trim = TRUE,
  ...
)

label_number_cz(
  accuracy = NULL,
  scale = 1,
  prefix = "",
  suffix = "",
  big.mark = " ",
  decimal.mark = ",",
  trim = TRUE,
  ...
)
}
\arguments{
\item{accuracy}{A number to round to. Use (e.g.) \code{0.01} to show 2 decimal
places of precision. If \code{NULL}, the default, uses a heuristic that should
ensure breaks have the minimum number of digits needed to show the
difference between adjacent values.

Applied to rescaled data.}

\item{scale}{A scaling factor: \code{x} will be multiplied by \code{scale} before
formatting. This is useful if the underlying data is very small or very
large.}

\item{prefix}{Additional text to display before the number. The suffix is
applied to absolute value before \code{style_positive} and \code{style_negative} are
processed so that \code{prefix = "$"} will yield (e.g.) \verb{-$1} and \verb{($1)}.}

\item{suffix}{Additional text to display after the number.}

\item{big.mark}{Character used between every 3 digits to separate thousands.}

\item{decimal.mark}{The character to be used to indicate the numeric
decimal point.}

\item{trim}{Logical, if \code{FALSE}, values are right-justified to a common
width (see \code{\link[base:format]{base::format()}}).}

\item{...}{Other arguments passed on to \code{\link[base:format]{base::format()}}.}
}
\description{
\code{label_percent_cz()} returns a formatter that outputs percent labels with a " \%" suffix and a decimal comma.
\code{label_number_cz()} uses space as thousand separator and decimal comma.
Use these when you need to format labels on something other than X and Y axes
}
\examples{
library(reschola)
library(ggplot2)
ggplot(mpg, aes(hwy, cty)) +
  geom_point(aes(colour = cty / max(cty), size = hwy * 100)) +
  theme_schola(family = "sans", title_family = "sans", gridlines = "scatter") +
  scale_color_binned(labels = label_percent_cz()) +
  scale_size_binned(labels = label_number_cz())
}
