% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive.R
\name{gd_backup}
\alias{gd_backup}
\title{Backup a file to project's Google Drive}
\usage{
gd_backup(file, dir = gd_get_proj())
}
\arguments{
\item{file}{asdf}

\item{dir}{asdf}
}
\value{

}
\description{
Backup a file to project's Google Drive
}
