% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive.R
\name{gd_download_folder}
\alias{gd_download_folder}
\title{Download files from Google Drive folder}
\usage{
gd_download_folder(
  folder_url,
  dest_dir = "data-input",
  files_from_subfolders = F,
  overwrite = F
)
}
\arguments{
\item{folder_url}{Folder URL.}

\item{dest_dir}{Local URL in which to store files.}

\item{files_from_subfolders}{Whether to download also files from subdirectories.
Defaults to \code{TRUE}. See Note.}

\item{overwrite}{Whether to overwrite if file of same name exists.}
}
\value{
vector of paths to downloaded files
}
\description{
Downloads all downloadable files from given Google Drive folder
and saves them to specified local directory. Files will have the same names.
Use \code{googledrive::drive_download()} for downloading individual files.
}
\note{
As it would be very difficult to mirror the folder exactly,
if \code{files_from_subfolders} is set to \code{TRUE}, files from subfolders will be
saved in the same directory, not in the respective subdirectories.
Bear this in mind when naming files in GDrive subfolders so as to avoid naming conflicts.
}
\examples{
\donttest{
gd_url <- "https://drive.google.com/drive/folders/1bCyR_VKAP_43NEujqisjN77hANnMKfHZ"
gd_download_folder(folder_url = gd_url,
                   files_from_subfolders = T, overwrite = T)
}
}
\seealso{
Other Workflow helpers: 
\code{\link{draft_redoc}()},
\code{\link{draft_word}()},
\code{\link{manage_docx_header_logos}()}
}
\concept{Workflow helpers}
