% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gd_get_proj}
\alias{gd_get_proj}
\title{Get current reschola project Google Drive URL ID}
\usage{
gd_get_proj(url_object = ".gd_proj_url")
}
\arguments{
\item{url_object}{\emph{character}, name of the object URL is stored in. \code{.gd_proj_url} by default.}
}
\value{
A character vector of class \code{drive_id}.
}
\description{
Gets a hidden object \code{.gd_proj_url} (by default) created at project
"startup". Use \code{usethis::edit_r_profile()} to change the URL.
}
