% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive.R
\name{gd_upload_file}
\alias{gd_upload_file}
\title{Upload a file to project's Google Drive}
\usage{
gd_upload_file(file, dir = gd_get_proj())
}
\arguments{
\item{file}{\emph{character}, path to the file.}

\item{dir}{Directory at Google Drive, defaults to current reschola project
directory picked up by \code{gd_get_proj()}.}
}
\value{
An object of class \code{dribble}, a tibble with one row per file.
}
\description{
Upload a file to project's Google Drive
}
