% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data}
\alias{get_data}
\title{Get RDS from custom dir}
\usage{
get_data(file, type, data_dir = "data")
}
\arguments{
\item{file}{file to make, rds fileext is appended automatically or forced if
other ext is provided}

\item{type}{type of data, dir inside data_dir}
}
\description{
Get RDS from custom dir
}
\keyword{internal}
