% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{get_lightness}
\alias{get_lightness}
\title{Extract perceived lightness from colour}
\usage{
get_lightness(col)
}
\arguments{
\item{col}{colour .item or hex}
}
\value{
L .resp from HSL-converted model
}
\description{
Extract perceived lightness from colour
}
\keyword{internal}
