% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rds}
\alias{get_rds}
\title{Get RDS}
\usage{
get_rds(file, type, data_dir = "data")
}
\arguments{
\item{file}{file to make, rds fileext is appended automatically or forced if
other ext is provided}

\item{type}{type of data, dir inside data_dir}
}
\description{
Get RDS
}
\keyword{internal}
