% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emailer.R
\name{gw_auth}
\alias{gw_auth}
\title{Authenticate to Google Workspace / Gmail}
\usage{
gw_auth(domain = "@scholaempirica.org")
}
\arguments{
\item{domain}{\emph{character}, email domain. Default to \@scholaempirica.org}
}
\value{
Called for side effects.
}
\description{
Obtains credentials and stores them in dedicated environment for further use by
\code{compose()} and \code{send()} functions.
}
\details{
The credentials used for user authentication are obtained through interactive
prompts, mainly for security reasons. The function tries to obtain the
credentials from the system environment variable table first. If none found,
the user is asked for them and is provided with a guidance for permanent
credential storage.
}
\examples{
\dontrun{
}

}
\concept{mailing functions}
