% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{install_reschola_fonts}
\alias{install_reschola_fonts}
\title{Install reschola fonts on your computer}
\usage{
install_reschola_fonts()
}
\value{
Side effects.
}
\description{
Install reschola fonts on your computer
}
