% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/czech_date.R
\name{knit_print.czech_date_interval}
\alias{knit_print.czech_date_interval}
\title{knit_print S3 method for class czech_date_interval}
\usage{
\method{knit_print}{czech_date_interval}(x, ...)
}
\description{
knit_print S3 method for class czech_date_interval
}
\keyword{internal}
