% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limesurvey.R
\name{ls_add_participants}
\alias{ls_add_participants}
\title{Add Participant(s) to the Survey}
\usage{
ls_add_participants(survey_id, part_data, create_token = TRUE)
}
\arguments{
\item{survey_id}{\emph{integer}, ID of the survey (as found with \code{ls_surveys},
e.g.).}

\item{part_data}{\emph{tibble / data.frame / list}, object with participant(s)
data, i.e., \code{firstname}, \code{lastname}, \code{email} etc.}

\item{create_token}{\emph{logical}, whether to create token outright. Defaults to
\code{TRUE}).}
}
\value{
Called for a side effect, but returns the inserted data including
additional new information like the token string.
}
\description{
The function takes a \code{tibble} (or any object that is internally represented
as a (named) \code{list} by \code{R}) of participant(s) data and adds them to the
LimeSurvey participant database of the selected survey.
}
\details{
Generally, your \code{part_data} object have to contain three variables:
\code{firstname}, \code{lastname}, and \code{email}. That is something like a bare minimum,
but you may add any attribute recognized by LimeSurvey -- even custom
attributes like \code{attribute_1} or so. For the human-readable list of custom
attributes being held in the LimeSurvey participant database of the selected
survey, use \code{ls_get_attrs()}. However, do not use the "semantic" form, as
\code{ls_add_participants()} recognizes only raw, i.e. \code{attribute_1} notation.
}
\examples{
\dontrun{
# create participant table
part_data <- tibble(
  firstname = "John",
  lastname = "Doe",
  email = "john@example.com",
  language = "cs",
  attribute_1 = "Example School"
)

# insert participant into the LimeSurvey database
ls_add_participants(123456, part_data)

# check if OK
ls_participants(123456)
}

}
\seealso{
Other LimeSurvey functions: 
\code{\link{ls_call}()},
\code{\link{ls_export}()},
\code{\link{ls_get_attrs}()},
\code{\link{ls_invite}()},
\code{\link{ls_login}()},
\code{\link{ls_participants}()},
\code{\link{ls_responses}()},
\code{\link{ls_set_participant_properties}()},
\code{\link{ls_surveys}()}
}
\concept{LimeSurvey functions}
