% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limesurvey.R
\name{ls_call}
\alias{ls_call}
\title{Call LimeSurvey API Directly}
\usage{
ls_call(method, params = list())
}
\arguments{
\item{method}{\emph{character}, a method supported by LimeSurvey API.}

\item{params}{\emph{list}, arguments of the method. Need to be in order stated in
documentation.}
}
\value{
A tibble, or raw object if server response cannot be reasonably
coerced to a tibble.
}
\description{
General function used internally by every other \code{ls_} fellow. Useful when you
want something special that is not (yet) implemented in \code{reschola}.
}
\details{
\href{https://api.limesurvey.org/classes/remotecontrol_handle.html}{The list of available LimeSurvey API calls} briefly
documents the functionality provided. You must strictly adhere to the
arguments positions, as they are passed as an array.
}
\examples{
\dontrun{
ls_call("get_survey_properties", params = list(iSurveyID = 123456))
}

}
\seealso{
Other LimeSurvey functions: 
\code{\link{ls_add_participants}()},
\code{\link{ls_export}()},
\code{\link{ls_invite}()},
\code{\link{ls_login}()},
\code{\link{ls_participants}()},
\code{\link{ls_responses}()},
\code{\link{ls_surveys}()}
}
\concept{LimeSurvey functions}
