% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limesurvey.R
\name{ls_check_attributes}
\alias{ls_check_attributes}
\title{Check for Illegal Attributes}
\usage{
ls_check_attributes(attributes)
}
\arguments{
\item{attributes}{vector of attributes}
}
\description{
Checks for illegal LS attributes and raises an error when any found.
}
