% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limesurvey.R
\name{ls_export}
\alias{ls_export}
\title{Export Responses with Participants Attached}
\usage{
ls_export(
  survey_id,
  attributes = TRUE,
  n_participants = 999,
  lang = "cs",
  part = "all",
  only_unused_tokens = FALSE
)
}
\arguments{
\item{survey_id}{\emph{integer}, ID of the survey (as found with \code{ls_surveys},
e.g.).}

\item{attributes}{\emph{character vector or list} not used at the moment,
LimeSurvey API error}

\item{n_participants}{\emph{integer}, the number of participants to list, defualt
to 999.}

\item{lang}{\emph{character}, ISO 639 language code, default to \code{cs}.}

\item{part}{\emph{character}, completion status, either \code{complete}, \code{incomplete}
or \code{all} (the default).}

\item{only_unused_tokens}{\emph{logical}, should only the unused tokens be listed?
Default to FALSE.}
}
\value{
A tibble.
}
\description{
Export Responses with Participants Attached
}
\examples{
\dontrun{
ls_export(123456)
}

}
\seealso{
Other LimeSurvey functions: 
\code{\link{ls_call}()},
\code{\link{ls_login}()},
\code{\link{ls_participants}()},
\code{\link{ls_responses}()},
\code{\link{ls_surveys}()}
}
\concept{LimeSurvey functions}
