% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limesurvey.R
\name{ls_get_attrs}
\alias{ls_get_attrs}
\title{Get Survey Attributes in Semantic Form}
\usage{
ls_get_attrs(survey_id)
}
\arguments{
\item{survey_id}{\emph{integer}, ID of the survey (as found, e.g., with
\code{ls_surveys()}).}
}
\value{
A character vector of "semantic" attributes with names denoting "raw"
attributes used internally by LimeSurvey.
}
\description{
A survey can comprise one or more custom attributes useful for encoding
participant characteristics directly within participants table. However,
LimeSurvey refers to them as \code{attribute_x} (where \code{x} is the attribute
position) which is not useful at all. The function aims to resolve this issue
by returning so-called "semantic" form of attributes with human-readable
description.
}
\examples{
\dontrun{
ls_get_attrs(123456)
}

}
\seealso{
Other LimeSurvey functions: 
\code{\link{ls_add_participants}()},
\code{\link{ls_call}()},
\code{\link{ls_export}()},
\code{\link{ls_invite}()},
\code{\link{ls_login}()},
\code{\link{ls_participants}()},
\code{\link{ls_responses}()},
\code{\link{ls_set_participant_properties}()},
\code{\link{ls_surveys}()}
}
\concept{LimeSurvey functions}
