% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limesurvey.R
\name{ls_invite}
\alias{ls_invite}
\title{Invite Participant(s)}
\usage{
ls_invite(survey_id, tid, uninvited_only = TRUE)
}
\arguments{
\item{survey_id}{\emph{integer}, ID of the survey (as found with \code{ls_surveys},
e.g.).}

\item{tid}{\emph{integer(s)}, one ore more token IDs (\strong{not tokens!}) from database
to invite.}

\item{b}{\emph{logical}, send invitation for participants that have not been
invited yet (default). If \code{FALSE}, send an invite even if already sent.}
}
\value{
Called for side-efect. Returns a message from the server.
}
\description{
Send an email with a link to a survey to the particular participant(s). Uses
email template specified in the LimeSurvey web interface.
}
\details{
LimeSurvey allows you to send so-called invitation to a participant, meaning
he or she will get an email containing a link with his or her unique access
token.
}
\examples{
\dontrun{
ls_invite(123456, 18)
}
}
