% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limesurvey.R
\name{ls_set_participant_properties}
\alias{ls_set_participant_properties}
\title{Set or Edit Attribute(s) of an Participant}
\usage{
ls_set_participant_properties(survey_id, tid, ...)
}
\arguments{
\item{survey_id}{\emph{integer}, ID of the survey (as found with \code{ls_surveys()},
e.g.).}

\item{...}{attributes in the form \code{attribute_name = attribute_value}.}

\item{participant}{\emph{integer} or \emph{list}, \strong{one} token ID (\strong{not token!}) from
participant database. Use \code{ls_participants()} to get the \code{tid}. Another
option is to pass a list of one ore more participant properties, i.e.
\code{list(lastname = "Doe")}}
}
\value{
A tibble with the participant row just edited.
}
\description{
Set or Edit Attribute(s) of an Participant
}
\examples{
\dontrun{
ls_set_participant_properties(123456,
  participant = 18, email = "new@email.cz",
  attribute_1 = 600123456
)
}

}
\seealso{
Other LimeSurvey functions: 
\code{\link{ls_add_participants}()},
\code{\link{ls_call}()},
\code{\link{ls_export}()},
\code{\link{ls_get_attrs}()},
\code{\link{ls_invite}()},
\code{\link{ls_login}()},
\code{\link{ls_participants}()},
\code{\link{ls_responses}()},
\code{\link{ls_surveys}()}
}
\concept{LimeSurvey functions}
