% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limesurvey.R
\name{ls_surveys}
\alias{ls_surveys}
\title{List All Surveys at the Server}
\usage{
ls_surveys()
}
\value{
A tibble.
}
\description{
Lists every single survey found at the server you are logged in. Useful to
gain information about the surveys IDs etc.
}
\examples{
\dontrun{
ls_surveys()
}

}
\seealso{
Other LimeSurvey functions: 
\code{\link{ls_add_participants}()},
\code{\link{ls_call}()},
\code{\link{ls_export}()},
\code{\link{ls_get_attrs}()},
\code{\link{ls_invite}()},
\code{\link{ls_login}()},
\code{\link{ls_participants}()},
\code{\link{ls_responses}()},
\code{\link{ls_set_participant_properties}()}
}
\concept{LimeSurvey functions}
