% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_rds}
\alias{make_rds}
\title{Make RDS}
\usage{
make_rds(.x, file, type, data_dir = "data")
}
\arguments{
\item{.x}{an object to save as RDS}

\item{file}{file to make, rds fileext is appended automatically or forced if
other ext is provided}

\item{type}{type of data, dir inside data_dir}

\item{data_dir}{dir with data dirs}
}
\description{
Make RDS
}
\keyword{internal}
