% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_header_logo.R
\name{manage_docx_header_logos}
\alias{manage_docx_header_logos}
\title{Add or replace logo in the header of a Word document}
\usage{
manage_docx_header_logos(
  docx_path,
  png_logo_path,
  action = c("add_client", "replace_schola"),
  height = NULL
)
}
\arguments{
\item{docx_path}{The Word document in which to replace logos. Must contain the bookmarks
\code{schola_logo} and \code{client_logo} in the header (files created from reschola templates do by default.)}

\item{png_logo_path}{a PNG file which will be added/used as replacement}

\item{action}{whether to add new logo on the right (\code{"add_client"}) or
replace default Schola logo on the left (\code{"replace_schola"})}

\item{height}{height of the new logo in the resulting document, in \code{cm}.
By default, uses the height of the existing Schola logo in the header.}
}
\value{
invisibly returns the name of the new Word doc, which is same as the input Word doc,
with an an added \verb{_addedlogo} suffix.
}
\description{
Takes an existing Word document created using the \code{reschola} templates
and adds (by default, or replaces if \code{action = "replace_schola"}) the logo with
the image file you point it to.
}
\note{
This requires specific bookmarks in the header of the input document.
This is taken from the skeleton.docx template in the template components.
If you overwrite the header in the input document, this function will not work.
}
\examples{
\dontrun{
library(reschola)
manage_docx_header_logos("draft.docx", png_logo_path = "logos/newlogo.png",
                         action = "add_client")
manage_docx_header_logos("draft.docx", png_logo_path = "logos/newlogo.png",
                         action = "replace_schola")
}
}
\seealso{
Other Workflow helpers: 
\code{\link{draft_pdf}()},
\code{\link{gd_download_folder}()}
}
\concept{Workflow helpers}
