% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{open_schola_template}
\alias{open_schola_template}
\title{Locate and open default Schola templates}
\usage{
open_schola_template(format = "pdf")
}
\arguments{
\item{format}{\emph{Character}, format which to look for. Defaults to \code{pdf}.}
}
\value{
No return value, called for side effect.
}
\description{
Open the default Schola templates used by \code{\link[=schola_word]{schola_word()}}, or
\code{\link[=schola_pdf]{schola_pdf()}} (specify with \code{format} argument). The templates are shipped
with the package and are somewhat hidden in \code{R} library, so this auxiliary
function can help you dig them up.
}
\details{
You can either edit the chosen template directly in its natural habitat
(questionable short-term solution), or better -- use "Save as" option and
keep it and use it within the project directory.
}
\examples{
\dontrun{
open_schola_template()
}
}
\seealso{
Other Report templates and formats: 
\code{\link{copy_schola_template}()},
\code{\link{schola_pdf}()},
\code{\link{schola_word2}()},
\code{\link{schola_word}()}
}
\author{
Jan Netik
}
\concept{Report templates and formats}
