% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{open_schola_word_template}
\alias{open_schola_word_template}
\title{Locate and open default Word template}
\usage{
open_schola_word_template()
}
\value{
No return value, called for side effect.
}
\description{
Open the default Word template used by \code{schola_word}.
The template comes with the package and is somewhat hidden in R library,
so this auxiliary function can help you dig it up.
}
\details{
You can either edit the template directly (see \link{schola_word2} for more detail) and save it, or better -
use "Save as" option and keep it within the project directory, as described in \link{schola_word2}.
}
\examples{
\dontrun{open_schola_word_template()}
}
\seealso{
Other Report templates and formats: 
\code{\link{schola_redoc}()},
\code{\link{schola_word2}()},
\code{\link{schola_word}()}
}
\author{
Jan Netik
}
\concept{Report templates and formats}
