% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lollipop.R
\name{plot_lollipop}
\alias{plot_lollipop}
\title{Plot lollipop}
\usage{
plot_lollipop(
  plot_data,
  direction = "blue_larger",
  var_labels = waiver(),
  negative_label = NULL,
  positive_label = NULL,
  ref_label = NULL,
  xlab = "rozdíl mezi Vaší školou a ostatními",
  observations_alpha = 0.2
)
}
\arguments{
\item{plot_data}{a}

\item{direction}{a}

\item{var_labels}{a}

\item{negative_label}{a}

\item{positive_label}{a}

\item{ref_label}{a}

\item{xlab}{a}

\item{observations_alpha}{opacity of individual observations}
}
\value{
ggplot2 plot
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
\seealso{
Other Making charts: 
\code{\link{flush_axis}},
\code{\link{prepare_lollipop_data}()},
\code{\link{schola_barplot}()},
\code{\link{theme_schola}()}
}
\concept{Making charts}
