% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.czech_date}
\alias{print.czech_date}
\title{Print Czech Date}
\usage{
\method{print}{czech_date}(date, ..., case = NULL)
}
\arguments{
\item{date}{\emph{date or date-like object} to parse.}

\item{case}{\emph{character}, either "nominative", "locative" or "genitive"
(default) or any unambiguous abbreviation of these.}
}
\description{
S3 method for class \code{czech_date}.
}
\keyword{internal}
