% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.czech_date_interval}
\alias{print.czech_date_interval}
\title{print S3 method for class czech_date_interval}
\usage{
\method{print}{czech_date_interval}(x, ...)
}
\description{
print S3 method for class czech_date_interval
}
\keyword{internal}
