% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R, R/scales.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{update_geom_font_defaults}
\alias{scale_x_comma}
\alias{scale_y_comma}
\alias{scale_x_percent}
\alias{scale_y_percent}
\alias{gg_check}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{hrbrthemes}{\code{\link[hrbrthemes]{gg_check}}, \code{\link[hrbrthemes]{scale_x_comma}}, \code{\link[hrbrthemes]{scale_x_percent}}, \code{\link[hrbrthemes]{scale_y_comma}}, \code{\link[hrbrthemes]{scale_y_percent}}, \code{\link[hrbrthemes]{update_geom_font_defaults}}}
}}

