% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{register_reschola_fonts}
\alias{register_reschola_fonts}
\title{Register \code{{reschola}} fonts on Windows}
\usage{
register_reschola_fonts(family = "Roboto Condensed")
}
\arguments{
\item{family}{font family to register, default is reschola recommended}
}
\value{
Called for side effects.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function is supposed to substitute \code{import_fonts()} as its
dependencies are a bit buggy today and not updated often. \strong{Not tested
properly yet.}
}
\details{
The function is only useful on Windows. On other platforms, fonts
should be available out of the box for \code{Cairo} and/or \code{AGG} devices (which
is recommended for PNG and other bitmap formats, see \code{{ragg}} package).
}
\seealso{
Other Font helpers and shortcuts: 
\code{\link{font_rc}},
\code{\link{import_fonts}()},
\code{\link{use_reschola_fonts}()}
}
\concept{Font helpers and shortcuts}
