% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_empty_at}
\alias{remove_empty_at}
\title{Remove rows empty only at specified columns}
\usage{
remove_empty_at(.data, vars)
}
\arguments{
\item{.data}{tibble or data.frame}

\item{vars}{variables to check for empty values, supports tidyselect syntax}
}
\value{
tibble or data.frame without observations with all \code{vars} empty
}
\description{
Drop rows that are completely empty \emph{at specified variables}. Note
\code{\link[tidyr:drop_na]{tidyr::drop_na()}} have a similar usage, but it drops rows containing \emph{any}
missing values (so only complete observations are retained), not \emph{all}
missings. \code{remove_empty_at()} drops only those rows that have no non-missing
data at the variables specified in \code{vars}.
}
\examples{
airquality |> remove_empty_at(c(Ozone, Solar.R))

}
