% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reschola-package.R
\docType{package}
\name{reschola-package}
\alias{reschola}
\alias{reschola-package}
\title{reschola: The Schola Empirica package}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A collection of utilies, themes and templates for data
    analysis at Schola Empirica.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/scholaempirica/reschola}
  \item Report bugs at \url{https://github.com/scholaempirica/reschola/issues}
}

}
\author{
\strong{Maintainer}: Petr Bouchal \email{pbouchal@gmail.com} (\href{https://orcid.org/0000-0002-0471-716X}{ORCID})

Other contributors:
\itemize{
  \item Jaromír Mazák [contributor]
  \item Aleš Vomáčka [contributor]
  \item Schola Empirica [copyright holder, funder]
}

}
\keyword{internal}
