% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\docType{data}
\name{font_rc}
\alias{font_rc}
\alias{font_rc_bold}
\alias{font_rc_thin}
\alias{font_rc_light}
\alias{font_r_bold}
\alias{font_r_thin}
\alias{font_r}
\alias{font_r_light}
\title{Roboto font shortcuts}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
font_rc

font_rc_bold

font_rc_thin

font_rc_light

font_r_bold

font_r_thin

font_r

font_r_light
}
\description{
\code{font_rc} == "\verb{Roboto Condensed}"

\code{font_rc_bold} == "\verb{Roboto Condensed Bold}"

\code{font_rc_thin} == "\verb{Roboto Condensed Thin}"

\code{font_rc_light} == "\verb{Roboto Condensed Light}"

\code{font_bold} == "\verb{Roboto Bold}"

\code{font_rc_thin} == "\verb{Roboto Thin}"

\code{font_r} == "\code{Roboto}"

\code{font_fc_light} == "\verb{Roboto Light}"
}
\note{
\code{font_rc_light} (a.k.a. "\verb{Roboto Condensed Light}") is not available on
Windows and will throw a warning if used in plots.

\code{font_r_light} (a.k.a. "\verb{Roboto Light}") is not available on
Windows and will throw a warning if used in plots.
}
\seealso{
Other Font helpers and shortcuts: 
\code{\link{import_fonts}()},
\code{\link{install_reschola_fonts}()},
\code{\link{register_reschola_fonts}()},
\code{\link{use_reschola_fonts}()}

Other Font helpers and shortcuts: 
\code{\link{import_fonts}()},
\code{\link{install_reschola_fonts}()},
\code{\link{register_reschola_fonts}()},
\code{\link{use_reschola_fonts}()}

Other Font helpers and shortcuts: 
\code{\link{import_fonts}()},
\code{\link{install_reschola_fonts}()},
\code{\link{register_reschola_fonts}()},
\code{\link{use_reschola_fonts}()}

Other Font helpers and shortcuts: 
\code{\link{import_fonts}()},
\code{\link{install_reschola_fonts}()},
\code{\link{register_reschola_fonts}()},
\code{\link{use_reschola_fonts}()}

Other Font helpers and shortcuts: 
\code{\link{import_fonts}()},
\code{\link{install_reschola_fonts}()},
\code{\link{register_reschola_fonts}()},
\code{\link{use_reschola_fonts}()}

Other Font helpers and shortcuts: 
\code{\link{import_fonts}()},
\code{\link{install_reschola_fonts}()},
\code{\link{register_reschola_fonts}()},
\code{\link{use_reschola_fonts}()}

Other Font helpers and shortcuts: 
\code{\link{import_fonts}()},
\code{\link{install_reschola_fonts}()},
\code{\link{register_reschola_fonts}()},
\code{\link{use_reschola_fonts}()}

Other Font helpers and shortcuts: 
\code{\link{import_fonts}()},
\code{\link{install_reschola_fonts}()},
\code{\link{register_reschola_fonts}()},
\code{\link{use_reschola_fonts}()}
}
\concept{Font helpers and shortcuts}
\keyword{datasets}
