% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{schola_barplot}
\alias{schola_barplot}
\title{Plot standard Schola likert-like barplot with groupwise comparison per items}
\usage{
schola_barplot(
  .data,
  vars,
  group,
  dict = dict_from_data(.data),
  escape_level = "nevím",
  n_breaks = 11,
  desc = TRUE,
  labels = TRUE,
  min_label_width = 0.09,
  absolute_counts = TRUE,
  fill_cols = NULL,
  fill_labels = waiver(),
  facet_label_wrap = 115,
  reverse = FALSE,
  order_by = "chi-square differences",
  drop = FALSE,
  drop_na = TRUE,
  show.legend = TRUE,
  ...
)
}
\arguments{
\item{.data}{data with items and group variable}

\item{vars}{vector of items, supports \code{{tidyselect}} syntax (i.e,
non-standard evaluation)}

\item{group}{group variable used to split the results, have to be logical,
where \code{TRUE} is gonna be considered as "focal" group and displayed as upper
group Supports \code{{tidyselect}} syntax.}

\item{dict}{item code-label dictionary, if none provided, those are derived
from the data.}

\item{escape_level}{\emph{character}, level of item response considered as NA}

\item{n_breaks}{number of breaks displayed at x-axis, outer labels are
automatically aligned to face inward. Defaults to 11, which results in 10\%
wide breaks.}

\item{desc}{sor items in descending order?}

\item{labels}{draw labels?}

\item{min_label_width}{smallest percentage (0-1) to display in the plot,
proportions larger than this value are shown, smaller are not.}

\item{absolute_counts}{draw labels and absolute counts in parentheses?}

\item{fill_cols}{colors to be used for item categories, defaults to NULL,
meaning standard RdYlBu palette will be used}

\item{fill_labels}{character vector or function taking breaks and returning
labels for fill aesthetic}

\item{facet_label_wrap}{width of facet label to wrap}

\item{reverse}{if TRUE, reverse colors}

\item{order_by}{how to order the items. chi-square differences (default)
computes chi-square test for every item and sort them by largest X2
statistic to smallest (if desc = TRUE)}

\item{drop}{Drop unobserved levels form the legend? Defaults to \code{FALSE}. See
\link[ggplot2:discrete_scale]{ggplot2::discrete_scale} for more details. To always show the legend key,
make sure you have \code{show.legend} set to \code{TRUE} as well (as done by
default).}

\item{drop_na}{Drop \code{NA}s from every item (a.k.a. "pairwise")? Defaults to
\code{TRUE}. Note that the number of observations per item may differ, because
\code{NA} in one item does not mean the respondent row is discarded completely
(listwise).}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{...}{
  Arguments passed on to \code{\link[=fct_nanify]{fct_nanify}}
  \describe{
    \item{\code{negate}}{\emph{logical}, whether to return non-matching elements. Defaults to
\code{FALSE}.}
    \item{\code{ignore_case}}{\emph{logical}, ignore case when matching? Defaults to \code{TRUE}.}
  }}
}
\value{
Object of class "gg", "ggplot".
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
\seealso{
Other Making charts: 
\code{\link{flush_axis}},
\code{\link{plot_lollipop}()},
\code{\link{prepare_lollipop_data}()},
\code{\link{theme_schola}()}
}
\concept{Making charts}
