% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{schola_labeller}
\alias{schola_labeller}
\title{Labeller with width wrapping and item code-label dictionary}
\usage{
schola_labeller(dict, width = 80)
}
\arguments{
\item{dict}{code-label dictionary as named character vector, see
\code{dict_from_data()}}

\item{width}{maximal string length per line, if \code{NULL}, this is off}
}
\value{
object of class \code{labeller}, to be used in \verb{facet_*} \code{{ggplot2}}
functions.
}
\description{
Turns item codes into labels by looking up provided dictionary and applies
them as facet labels.
}
