% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_output.R
\name{schola_pdf}
\alias{schola_pdf}
\title{Schola Empirica PDF Format}
\usage{
schola_pdf(
  num_format = NULL,
  fig_crop = TRUE,
  number_sections = FALSE,
  toc = TRUE,
  template = find_resource("schola_pdf", "schola_template.tex"),
  latex_engine = "xelatex",
  document_class = "report",
  ...
)
}
\arguments{
\item{num_format}{\emph{Character}, if \code{cs}, Czech number formatting is used.
Other values are currently ignored and will result in default options.}

\item{fig_crop}{\emph{Logical}, whether to crop transparent or white margins of
figures. Defaults to \code{TRUE}. Note that for cropping to work, two
dependencies needs to be satisfied, see \code{?ensure_cropping}.}

\item{number_sections}{\emph{Logical}, \code{TRUE} to number headings. Defaults to
\code{FALSE}.}

\item{toc}{\emph{Logical}, \code{TRUE} (default) to include a table of contents. The
title is set with \code{pandoc} variable \code{toc-title} in YAML header.}

\item{template}{\emph{Character}, path to the \code{.tex} template used by the format.
Defaults to standard Schola template bundled in the package. \strong{Changes
discouraged.}}

\item{latex_engine}{\emph{Character}, engine used for \code{.md} to \code{.pdf} conversion.
Only \code{xelatex} (the default) and \code{lualatex} are supported because of custom
fonts. \strong{Changes discouraged.}}

\item{document_class}{\emph{Character}, one of standard LaTeX document class.
Defaults to \code{report}. \strong{Changes discouraged.}}

\item{...}{
  Arguments passed on to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}
  \describe{
    \item{\code{toc_depth}}{Depth of headers to include in table of contents}
    \item{\code{fig_width}}{Default width (in inches) for figures}
    \item{\code{fig_caption}}{\code{TRUE} to render figures with captions}
    \item{\code{df_print}}{Method to be used for printing data frames. Valid values
include "default", "kable", "tibble", and "paged". The "default" method
uses a corresponding S3 method of \code{print}, typically
\code{print.data.frame}. The "kable" method uses the
\code{\link[knitr:kable]{knitr::kable}} function. The "tibble" method uses
the \pkg{tibble} package to print a summary of the data frame. The "paged"
method creates a paginated HTML table (note that this method is only valid
for formats that produce HTML). In addition to the named methods you can
also pass an arbitrary function to be used for printing data frames. You
can disable the \code{df_print} behavior entirely by setting the option
\code{rmarkdown.df_print} to \code{FALSE}. See
\href{https://bookdown.org/yihui/rmarkdown/html-document.html#data-frame-printing}{Data
frame printing section} in bookdown book for examples.}
    \item{\code{highlight}}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso",
"zenburn", and "haddock". Pass \code{NULL} to prevent syntax highlighting.}
    \item{\code{keep_tex}}{Keep the intermediate tex file used in the conversion to PDF}
    \item{\code{keep_md}}{Keep the markdown file generated by knitting.}
    \item{\code{citation_package}}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{default} if neither package is to be used,
which means citations will be processed via the command
\command{pandoc-citeproc}.}
    \item{\code{includes}}{Named list of additional content to include within the
document (typically created using the \code{\link[rmarkdown]{includes}} function).}
    \item{\code{md_extensions}}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link[rmarkdown]{rmarkdown_format}} for
additional details.}
    \item{\code{output_extensions}}{Pandoc extensions to be added or removed from the
output format, e.g., \code{"-smart"} means the output format will be
\code{latex-smart}.}
    \item{\code{pandoc_args}}{Additional command line options to pass to pandoc}
    \item{\code{extra_dependencies}}{A LaTeX dependency \code{latex_dependency()}, a
list of LaTeX dependencies, a character vector of LaTeX package names (e.g.
\code{c("framed", "hyperref")}), or a named list of LaTeX package options
with the names being package names (e.g. \code{list(hyperef =
c("unicode=true", "breaklinks=true"), lmodern = NULL)}). It can be used to
add custom LaTeX packages to the .tex header.}
  }}
}
\value{
A modified \code{pdf_document2} with the standard Schola formatting.
}
\description{
Function called in the output section of the YAML header (see examples
below), instructing \code{knitr} to use the standard Schola PDF document format.
}
\examples{
\dontrun{
output:
  reschola::schola_pdf:
    toc:false
}

}
\seealso{
Other Report templates and formats: 
\code{\link{open_schola_template}()},
\code{\link{schola_redoc}()},
\code{\link{schola_word2}()},
\code{\link{schola_word}()}
}
\author{
Jan Netik
}
\concept{Report templates and formats}
