% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_project_bindings.R
\name{schola_project}
\alias{schola_project}
\title{FUNCTION_TITLE}
\usage{
schola_project(path, ...)
}
\arguments{
\item{DESCRIPTION.}{}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
