% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_output.R
\name{schola_redoc}
\alias{schola_redoc}
\title{Reversible Schola Empirica word document}
\usage{
schola_redoc(...)
}
\arguments{
\item{...}{Arguments to be passed to \verb{[redoc::redoc]} and secondarily to
\code{word_document}.}
}
\value{
A modified \code{redoc} with the standard Schola formatting.
}
\description{
This is a function called in the output of the yaml of the Rmd file to
provide "reversible markdown-Word collaboration" as devised by \code{redoc},
while using the standard Schola word document formatting.
See the \href{https://noamross.github.io/redoc/}{redoc documentation} for a how to
on the reversible editing.
}
\details{
The capabilities of this format are a bit more limited compared to \code{schola_word}.
}
\examples{
\dontrun{
output:reschola::schola_redoc
}
}
\seealso{
Other Report templates and formats: 
\code{\link{open_schola_word_template}()},
\code{\link{schola_word2}()},
\code{\link{schola_word}()}
}
\concept{Report templates and formats}
