% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_output.R
\name{schola_word2}
\alias{schola_word2}
\title{Schola Empirica Word document with customisable template}
\usage{
schola_word2(...)
}
\arguments{
\item{...}{Arguments to be passed to \verb{[bookdown::word_document2]}}
}
\value{
A modified \code{word_document2} with the standard Schola formatting, but
without hard-coded and unchangeable template.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This is a function called in the output part of the YAML section of the Rmd
file while using the Word template provided at the same place (see example
below).
}
\details{
Compared to \code{schola_word}, this "version" comes with no predefined
template, so the user can utilize a template stated in YAML header (see the
example below, or read the \href{https://bookdown.org/yihui/rmarkdown-cookbook/word-template.html}{\code{bookdown} manual} for
more details and for a brief guide to Word templating).
}
\examples{
\dontrun{
output:
reschola::schola_word2:
reference_docx:template.docx
}
}
\seealso{
Other Report templates and formats: 
\code{\link{copy_schola_template}()},
\code{\link{open_schola_template}()},
\code{\link{schola_pdf}()},
\code{\link{schola_word}()}
}
\author{
Jan Netik
}
\concept{Report templates and formats}
