% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello, world!}
\usage{
hello()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
This is an example function named 'hello'
which prints 'Hello, world!'.

You can learn more about package authoring with RStudio at:

http://r-pkgs.had.co.nz/

Some useful keyboard shortcuts for package authoring:

Install Package:           'Ctrl + Shift + B'
Check Package:             'Ctrl + Shift + E'
Test Package:              'Ctrl + Shift + T'
}
