% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{use_reschola_fonts}
\alias{use_reschola_fonts}
\alias{set_reschola_ggplot_fonts}
\title{Make ggplot2 use font(s) in text-based geoms}
\usage{
use_reschola_fonts(
  family = "Ubuntu Condensed",
  face = "plain",
  size = 3.5,
  color = "#2b2b2b"
)

set_reschola_ggplot_fonts(
  family = "Ubuntu Condensed",
  face = "plain",
  size = 3.5,
  color = "#2b2b2b"
)
}
\arguments{
\item{family}{font family, defaults to reschola recommended}

\item{face}{font face}

\item{size}{font size}

\item{color}{font color}
}
\description{
Sets \code{{ggplot2}} defaults for most text-based geoms in \code{{ggplot2}},
\code{{ggtext}}, and \code{{ggrepel}}.
}
\details{
Geoms covered: "text", "label", "richtext", "text_box", "text_repel", and
"label_repel".
}
\seealso{
Other Font helpers and shortcuts: 
\code{\link{import_fonts}()},
\code{\link{install_reschola_fonts}()},
\code{\link{register_reschola_fonts}()}
}
\concept{Font helpers and shortcuts}
