% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_data}
\alias{write_data}
\title{Write RDS to custom dir}
\usage{
write_data(.data, file, type, data_dir = "data")
}
\arguments{
\item{.data}{an object to save as RDS}

\item{file}{file to make, rds fileext is appended automatically or forced if
other ext is provided}

\item{type}{type of data, dir inside data_dir}

\item{data_dir}{dir with data dirs}
}
\description{
Write RDS to custom dir
}
\keyword{internal}
