% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gd_get_proj}
\alias{gd_get_proj}
\title{Get current reschola project Google Drive URL ID}
\usage{
gd_get_proj(url_object = ".gd_proj_url")
}
\arguments{
\item{url_object}{\emph{character}, name of the object URL is stored in.
\code{.gd_proj_url} by default.}
}
\value{
A character vector of class \code{drive_id}.
}
\description{
Gets a hidden object \code{.gd_proj_url} (by default) created at project
"startup". See the details below for reschola projects created prior reschola
version 0.4.0.
}
\section{Legacy usage prior to \{reschola\} version 0.4.0}{

Call \code{usethis::edit_r_profile(scope = "project")} and write (note the dot
prefix):\preformatted{.gd_proj_url <- "your_google_drive_url"
}

Note that \strong{you have to restart your R session to apply any changes}. Note
also that \strong{the URL cannot contain any query} (parts with a leading
question mark), i.e. \code{?usp=sharing}.
}

