% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive.R
\name{gd_upload_file}
\alias{gd_upload_file}
\title{Upload a file to a Google Drive folder}
\usage{
gd_upload_file(file, target_folder = gd_url)
}
\arguments{
\item{file}{path to file}

\item{target_folder}{Google Drive Folder URL, defaults to gd_url}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
