% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{import_fonts}
\alias{import_fonts}
\title{Import fonts needed for theme_schola defaults}
\usage{
import_fonts()
}
\description{
Imports Roboto and Roboto Condensed. Draws on the hrbrthemes package.
}
\seealso{
Other Font helpers and shortcuts: 
\code{\link{font_rc}},
\code{\link{import_roboto}()},
\code{\link{set_reschola_ggplot_fonts}()}
}
\concept{Font helpers and shortcuts}
