% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{import_fonts}
\alias{import_fonts}
\title{Import Roboto fonts for use in charts and in the PDF reports}
\usage{
import_fonts()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#questioning}{\figure{lifecycle-questioning.svg}{options: alt='[Questioning]'}}}{\strong{[Questioning]}}

The function relies on \code{{extrafont}}
package that comes with briken dependencies at the moment. An experimental
\code{register_reschola_fonts()} is proposed.
}
\details{
This is an analogue of \code{hrbrthemes::import_roboto_condensed()}.

There is an option \code{reschola.loadfonts} which -- if set to \code{TRUE} -- will
call \code{extrafont::loadfonts()} to register non-core fonts with R PDF &
PostScript devices. If you are running under Windows, the package calls the
same function to register non-core fonts with the Windows graphics device.
}
\note{
If you install the fonts just for the current user (via right-click and
Install), they will probably \strong{not be discoverable} by the \code{fontspec}
LaTeX package that is used for PDF report typesetting!
}
\section{Credits}{
 Roboto is a trademark of Google.
}

\seealso{
Other Font helpers and shortcuts: 
\code{\link{font_rc}},
\code{\link{register_reschola_fonts}()},
\code{\link{set_reschola_ggplot_fonts}()}
}
\concept{Font helpers and shortcuts}
