% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/czech_date.R
\name{knit_print.czech_date}
\alias{knit_print.czech_date}
\title{knit_print S3 method for class czech_date}
\usage{
\method{knit_print}{czech_date}(x, ...)
}
\description{
knit_print S3 method for class czech_date
}
\keyword{internal}
