% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limesurvey.R
\name{ls_export}
\alias{ls_export}
\title{Export Responses with Participants Attached}
\usage{
ls_export(
  survey_id,
  attributes = TRUE,
  n_participants = 999,
  lang = "cs",
  part = "all",
  only_unused_tokens = FALSE,
  join_by = "token",
  ...
)
}
\arguments{
\item{survey_id}{\emph{integer}, ID of the survey (as found with \code{ls_surveys},
e.g.).}

\item{attributes}{\emph{logical} try to recover all attributes (default to TRUE),
or \emph{character vector} specifying requested attributes.}

\item{n_participants}{\emph{integer}, the number of participants to list, defualt
to 999.}

\item{lang}{\emph{character}, ISO 639 language code, default to \code{cs}.}

\item{part}{\emph{character}, completion status, either \code{complete}, \code{incomplete}
or \code{all} (the default).}

\item{only_unused_tokens}{\emph{logical}, should only the unused tokens be listed?
Default to FALSE.}

\item{join_by}{\emph{character}, the joining variable present in both responses
and participants tibbles. Default to \code{token}. Pass \code{NULL} to join by common
variables.}

\item{...}{
  Arguments passed on to \code{\link[dplyr:mutate-joins]{dplyr::left_join}}
  \describe{
    \item{\code{copy}}{If \code{x} and \code{y} are not from the same data source,
and \code{copy} is \code{TRUE}, then \code{y} will be copied into the
same src as \code{x}.  This allows you to join tables across srcs, but
it is a potentially expensive operation so you must opt into it.}
    \item{\code{suffix}}{If there are non-joined duplicate variables in \code{x} and
\code{y}, these suffixes will be added to the output to disambiguate them.
Should be a character vector of length 2.}
    \item{\code{keep}}{Should the join keys from both \code{x} and \code{y} be preserved in the
output?}
  }}
}
\value{
A tibble.
}
\description{
Export Responses with Participants Attached
}
\examples{
\dontrun{
ls_export(123456)
}

}
\seealso{
Other LimeSurvey functions: 
\code{\link{ls_add_participants}()},
\code{\link{ls_call}()},
\code{\link{ls_get_attrs}()},
\code{\link{ls_invite}()},
\code{\link{ls_login}()},
\code{\link{ls_participants}()},
\code{\link{ls_responses}()},
\code{\link{ls_set_participant_properties}()},
\code{\link{ls_surveys}()}
}
\concept{LimeSurvey functions}
