% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limesurvey.R
\name{ls_participants}
\alias{ls_participants}
\title{List Participants}
\usage{
ls_participants(
  survey_id,
  attributes = TRUE,
  n_participants = 999,
  only_unused_tokens = FALSE,
  translate_attrs = TRUE
)
}
\arguments{
\item{survey_id}{\emph{integer}, ID of the survey (as found, e.g., with
\code{ls_surveys()}).}

\item{attributes}{\emph{logical} try to recover all attributes (default to TRUE),
or \emph{character vector} specifying requested attributes.}

\item{n_participants}{\emph{integer}, the number of participants to list, defualt
to 999.}

\item{only_unused_tokens}{\emph{logical}, should only the unused tokens be listed?
Default to FALSE.}

\item{translate_attrs}{\emph{logical}, should the custom attributes be
"translated" to "semantic" version? Default to TRUE.}
}
\value{
A tibble, or raw object if server response cannot be reasonably
coerced to a tibble.
}
\description{
Fetches participant list of the solicited survey. The function tries to
retrieve as many attributes as possible and translate them to their
"semantic" version by default. You can also provide character vector of
requested attributes, but not in the "semantic" form (use, e.g.,
\code{attribute_1} or \code{usesleft}).
}
\examples{
\dontrun{
ls_participants(123456, attributes = c("usesleft"))
}

}
\seealso{
Other LimeSurvey functions: 
\code{\link{ls_add_participants}()},
\code{\link{ls_call}()},
\code{\link{ls_export}()},
\code{\link{ls_get_attrs}()},
\code{\link{ls_invite}()},
\code{\link{ls_login}()},
\code{\link{ls_responses}()},
\code{\link{ls_surveys}()}
}
\concept{LimeSurvey functions}
