% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lollipop.R
\name{plot_lollipop}
\alias{plot_lollipop}
\title{Title}
\usage{
plot_lollipop(
  plot_data,
  direction = 1,
  var_labels = waiver(),
  negative_label = NULL,
  positive_label = NULL,
  ref_label = NULL
)
}
\arguments{
\item{plot_data}{a}

\item{direction}{a}

\item{var_labels}{a}

\item{negative_label}{a}

\item{positive_label}{a}

\item{ref_label}{a}
}
\value{

}
\description{
Title
}
