% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lollipop.R
\name{prepare_lollipop_data}
\alias{prepare_lollipop_data}
\title{Prepare data for \code{plot_lollipop()}}
\usage{
prepare_lollipop_data(.data, vars, group)
}
\arguments{
\item{.data}{tibble or data.frame with variables}

\item{vars}{variables to reshape from wide to long, uses tidyselect syntax}

\item{group}{grouping variable (have to be logical!), usually denoting for
which school the plot should be tailored for}
}
\value{
several tibbles inside list, intended for data-mining and for
plotting
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
\seealso{
Other Making charts: 
\code{\link{flush_axis}},
\code{\link{plot_lollipop}()},
\code{\link{schola_barplot}()},
\code{\link{theme_schola}()}
}
\concept{Making charts}
