% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lollipop.R
\name{prepare_lollipop_data}
\alias{prepare_lollipop_data}
\title{Title}
\usage{
prepare_lollipop_data(.data, vars, group)
}
\arguments{
\item{.data}{a}

\item{vars}{a}

\item{group}{a}
}
\value{

}
\description{
Title
}
