% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{font_rc}
\alias{font_rc_light}
\alias{update_geom_font_defaults}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{hrbrthemes}{\code{\link[hrbrthemes]{font_rc}}, \code{\link[hrbrthemes]{font_rc_light}}, \code{\link[hrbrthemes]{update_geom_font_defaults}}}
}}

