% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{register_reschola_fonts}
\alias{register_reschola_fonts}
\title{Register reschola fonts on Windows}
\usage{
register_reschola_fonts(family = c("Ubuntu", "Ubuntu Condensed"))
}
\arguments{
\item{family}{font family/families to register, default is reschola
recommended.}
}
\value{
Called for side effects, but returns logical on process result
invisibly.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

The function is only for Windows, where it tries to register font family
provided with "Windows bitmap device". On other platforms, fonts \emph{should}
be readily available after installation. Note that even on Windows, you can
instruct RStudio to use smarter graphics device, such as "AGG" or "Cairo"
(\emph{Tools > Global Options > General > Graphics > Backend}).
}
\seealso{
Other Font helpers and shortcuts: 
\code{\link{import_fonts}()},
\code{\link{install_reschola_fonts}()},
\code{\link{use_reschola_fonts}()}
}
\concept{Font helpers and shortcuts}
