% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{schola_barplot}
\alias{schola_barplot}
\title{Plot standard Schola likert-like barplot with groupwise comparison per items}
\usage{
schola_barplot(
  .data,
  vars,
  group,
  dict = dict_from_data(.data),
  escape_level = "nevím",
  n_breaks = 11,
  desc = TRUE,
  labels = TRUE,
  min_label_width = 0.09,
  absolute_counts = TRUE,
  fill_labels = waiver(),
  reverse = FALSE,
  order_by = "chi-square differences",
  ...
)
}
\arguments{
\item{.data}{data with items and group variable}

\item{vars}{vector of items, supports {tidyselect} syntax (i.e, non-standard
evaluation)}

\item{group}{group variable used to split the results, have to be logical,
where \code{TRUE} is gonna be considered as "focal" group and displayed as upper
group with full opacity. Supports {tidyselect} syntax.}

\item{dict}{item code-label dictionary, if none provided, those are derived
from the data.}

\item{escape_level}{\emph{character}, level of item response considered as NA}

\item{n_breaks}{number of breaks displayed at x-axis, outer labels are
automatically aligned to face inward. Defaults to 11, which results in 10\%
wide breaks.}

\item{desc}{sor items in descending order?}

\item{labels}{draw labels?}

\item{min_label_width}{smallest percentage (0-1) to display in the plot,
proportions larger than this value are shown, smaller are not.}

\item{absolute_counts}{draw labels and absolute counts in parentheses?}

\item{fill_labels}{character vector or function taking breaks and returning
labels for fill aesthetic}

\item{reverse}{if TRUE, reverse colors}

\item{order_by}{how to order the items. chi-square differences (default)
computes chi-square test for every item and sort them by largest X2
statistic to smallest (if desc = TRUE)}

\item{...}{
  Arguments passed on to \code{\link[=fct_nanify]{fct_nanify}}
  \describe{
    \item{\code{negate}}{\emph{logical}, whether to return non-matching elements. Defaults to
\code{FALSE}.}
    \item{\code{ignore_case}}{\emph{logical}, ignore case when matching? Defaults to \code{TRUE}.}
  }}
}
\value{
object of class "gg", "ggplot"
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

TODO description...
}
\details{
TODO details...
}
