% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_output.R
\name{schola_pdf}
\alias{schola_pdf}
\title{Basic Schola Empirica PDF document}
\usage{
schola_pdf(num_format = "cs", ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[bookdown:html_document2]{bookdown::pdf_document2}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
A modified \code{pdf_document2} with the standard Schola formatting.
}
\description{
This is a function called in the output of the YAML of the Rmd file to
specify using the standard Schola PDF document formatting.
}
\examples{
\dontrun{
output:
  reschola::schola_pdf:
    toc: no
}
}
\seealso{
Other Report templates and formats: 
\code{\link{open_schola_word_template}()},
\code{\link{schola_redoc}()},
\code{\link{schola_word2}()},
\code{\link{schola_word}()}
}
\author{
Jan Netik
}
\concept{Report templates and formats}
