% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_project_bindings.R
\name{schola_project}
\alias{schola_project}
\title{schola_project}
\usage{
schola_project(path, ...)
}
\arguments{
\item{path}{path}

\item{...}{other params from new project dialog, created in DCF file}
}
\value{
TRUE
}
\description{
Function to parse inputs from New project dialog and create new project
}
\keyword{internal}
