% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_input_data}
\alias{get_input_data}
\alias{get_intermediate_data}
\alias{get_processed_data}
\alias{write_input_data}
\alias{write_intermediate_data}
\alias{write_processed_data}
\title{Quick access to data instandard \code{{reschola}} project}
\usage{
get_input_data(file)

get_intermediate_data(file)

get_processed_data(file)

write_input_data(.data, file)

write_intermediate_data(.data, file)

write_processed_data(.data, file)
}
\arguments{
\item{file}{\emph{character}, file name to get or write to, \code{.rds} is appended
automatically if not already provided.}

\item{.data}{\emph{data object} to save as RDS.}
}
\description{
Note that file paths are handled with \code{here()} already.
}
