% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_output.R
\name{schola_word}
\alias{schola_word}
\title{Basic Schola Empirica word document}
\usage{
schola_word(...)
}
\arguments{
\item{...}{Arguments to be passed to \verb{[bookdown::word_document2]}}
}
\value{
A modified \code{word_document2} with the standard Schola formatting.
}
\description{
This is a function called in the output of the yaml of the Rmd file to
specify using the standard Schola word document formatting.
}
\examples{
\dontrun{
  output: reschola::schola_word
}
}
