% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sten}
\alias{sten}
\title{Transform to STEN score (Standard Ten)}
\usage{
sten(score, standardize = FALSE, bounds = TRUE)
}
\arguments{
\item{score}{\emph{numeric} vector of scores to transform}

\item{standardize}{\emph{logical}, center and scale \code{score} vector before the
transformation? Defaults to \code{FALSE}.}

\item{bounds}{limit result to 1-10 scale, TRUE by default}
}
\value{
\emph{numeric} vector of transformed scores
}
\description{
Get your raw total score transformed to stens, i.e., with the mean of 5.5 and
a standard deviation of 2. See the \href{https://en.wikipedia.org/wiki/Sten_scores}{Wikipedia article} on the topic for more
details.
}
\examples{
rnorm(10) |> sten()
}
