% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_schola}
\alias{theme_schola}
\title{A Schola Empirica ggplot2 theme}
\usage{
theme_schola(
  base_size = 11,
  gridlines = c("y", "x", "both", "scatter"),
  family = "sans",
  title_family,
  side_margin = 6,
  plot.title.position = "plot",
  axis.title = ggplot2::element_blank(),
  ...
)
}
\arguments{
\item{base_size}{DESCRIPTION.}

\item{gridlines}{DESCRIPTION.}

\item{family}{DESCRIPTION.}

\item{title_family}{DESCRIPTION.}

\item{side_margin}{DESCRIPTION.}

\item{plot.title.position}{DESCRIPTION.}

\item{axis.title}{DESCRIPTION.}

\item{...}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
\seealso{
Other Making charts: 
\code{\link{flush_axis}}
}
\concept{Making charts}
