% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{use_reschola_fonts}
\alias{use_reschola_fonts}
\alias{set_reschola_ggplot_fonts}
\title{Make {ggplot2} use chosen font in geom_text/label}
\usage{
use_reschola_fonts(
  family = "Roboto Condensed",
  face = "plain",
  size = 3.5,
  color = "#2b2b2b"
)

set_reschola_ggplot_fonts(
  family = "Roboto Condensed",
  face = "plain",
  size = 3.5,
  color = "#2b2b2b"
)
}
\arguments{
\item{family}{font family}

\item{face}{font face}

\item{size}{font size}

\item{color}{font color}
}
\description{
Sets ggplot2 defaults for most text geoms from ggplot2, ggtext, and ggrepel.
}
\details{
Geoms covered: "text", "label", "richtext", "text_box", "text_repel", "label_repel".
}
\seealso{
Other Font helpers and shortcuts: 
\code{\link{font_rc}},
\code{\link{import_fonts}()},
\code{\link{register_reschola_fonts}()}
}
\concept{Font helpers and shortcuts}
